pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trache"
version = "0.1.0"
description = "Local-first Trello cache with Git-style sync — optimised for on-machine AI workflows"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "typer>=0.9",
    "httpx>=0.25",
    "pydantic>=2.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
oauth = ["authlib>=1.3"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]

[project.scripts]
trache = "trache.cli.app:app"

[tool.hatch.build.targets.wheel]
packages = ["src/trache"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
